requirements.txt
anyio==4.2.0
appdirs==1.4.4
beautifulsoup4==4.12.3
bs4==0.0.2
cached-property==1.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
commonmark==0.9.1
dictdiffer==0.9.0
exceptiongroup==1.2.0
filedate==3.0
fs==2.4.16
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
notion==0.0.28
notion-client==2.2.1
pillow==10.2.0
progressbar2==4.3.2
python-dateutil==2.8.2
python-frontmatter==1.1.0
python-slugify==8.0.4
python-utils==3.8.2
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
text-unidecode==1.3
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.2.0
